ensure_list even if we pass undefined
When the Annotated Transcript was generated, both
statistics and playback have shown "react stacktrace".
For some reason `data.transcript.head.property` was empty
on the data I've used for testing. That's probably due to
autodetection failing on the header (not sure if I have a
supported transcript format).
Nevertheless the UI should handle such cases.